backend/requirements.txt
Django
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
